bigspeak_scraper/module_1/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.6.1
python-dotenv==1.0.0
lxml==4.9.3